add parsed statements to current statement group
statement groups are yields when SET STATEMENT_GROUP is encounted; you
will need to retain the obj and read annotations/statements only when
UnsetStatementGroup is finally seen.  this will ensure that it is read
after the annotations and statements are visited.

regenerated ragel FSM with - ragel -F0 -L -R bel.rl