include exported function definitions file
The bel.rb C extension is non standard in the sense that loads libbel
using FFI. In order to export libbel functions a custom definitions
file, libbel.def, was added. The mkmf library expects the definition
file to be named as "${extension}.def" and located in $srcdir which
it where the C extension source it located (ext/mri).

This allows cross compilation to mingw to export the specified
functions described in libbel.def.