correct parse of STRING fsm definition
The TLDR is use the Longest-match Kleen star (e.g. **) when you want to
match the longest sequence within a machine before exiting to the
subsequent expression.

Detail:
Previously the STRING fsm definition used the Kleen star (*) to match
either a non-quote or an escape quote. This definition is sufficient for
STRING since a terminal quote will complete the token, but this is not
the case in a LIST (e.g. { "val1", "val2" }). For LIST we need to use
the longest-match Kleen star (**) to stay within the machine and match
the longest sequence before moving to the subsequent expression. In LIST
we could possibly move to a '}' or ',' which we were doing in error.

Notes:
- Added spec test to cover commas and escaped quotes in LISTs.

closes #37