moved RDF functionality to extension
This change removes BEL::Features as a mechanism for dynamic loading of
the RDF functionality. Instead it now is loaded as a Format extension
that also performs OpenClass on BEL::Language and BEL::Namespace
classes (e.g. NamespaceDefinition, Parameter, Term, and Statement).

It is worth noting that removal of BEL::Features may be a breaking
change so it is something to note in the 0.3.0 release. A deprecation
warning is most likely not useful as this module would/should have never
been used.