allow whitespace (space/tab) in BEL script records
This change required mutation of the document_main machine
instantiation. Specifically I needed to add SP* at the beginning of each
matched line and use left-guarded concatentation to place higher
priority on the whitespace expression. The left-guarded concatentation
is useful when you want to prefix an expression and both the left and
right expression have common characters.

added integration test to parse BEL with heavy use of space/tab
whitespace.

closes #13
closes #25