exclude Libbel spec tests on JRuby
I could not figure out how to use RSpec's skip call since it needs to be
within a describe bloc. The Libbel code needs to be prevented from being
loaded because the C extension does not exist on JRuby.

Providing an RSpec configure block with some exclusions might be better,
but it seems to heavy-handed at the moment.