complete ns values using Namespaces API
This change allows the Completion API to use the Namespaces API which
leverages the RDF namespace data. This fixed issues where bel.rb
contained incorrect hardcoded URIs for openbel namespaces.

When searching namespace values, without a prefix, the corresponding
prefix will be returned. For example a string of "p(AKT" will yield a
completion of "p(HGNC:AKT1" containing the prefix of the corresponding
namespace scheme.