move requiring of plugin deps on create
This moves requiring plugin dependencies to when the create method is
called on plugins.

Positive implications:
- Does not require libraries that may never be used. Reduces memory
  footprint.
- Does not show plugin load errors.

Negative implications:
- No insight into whether the plugin has all of its dependencies.

The ruby way would be to address testing plugin availability would be
to require it and let it fail which is the effect of this change. A more
useful experience for the user would be to test plugin availability and
report as part of "bel plugins --list".

refs #95