add rdf:type statements for parameter encoding
This changes BEL::Model::Parameter to attempt to load namespace data
(BELNS URL) and obtain the actual encoding (e.g. GRP, B, etc.). If any
error occurs a ruby warning is logged to STDERR (can suppress with -W0)
and the encoding is set as nil.

This solution works on the .belns URLs of namespaces. In the future RDF
URIs will allow us to fetch type information directly from the
NamespaceConcept RDF resource.

closes #96