write RDF statements directly to IO
This mechanism supports the added RDF translators, but is not ideal.
JSON-LD, TriG, RDF/XML, and RDFa require the set of statements before
serializing otherwise the output is malformed. This can be achieved by
separating into two RDF writer modules that we can choose for either
streaming or whole-graph serialization.

refs #95