support writing of BEL RDFS schema for translators
Re-enabled use of --schema / --no-schema in the bel2rdf.rb script.

Removed deprecation of bel2rdf.rb. This script is useful as it allows an
assumption of RDF output format that "bel translate" cannot provide.
This is useful for setting specific options. In the future this
executable will likely move to a bel-rdf module along with the set of
translators.

Updated each RDF translator to take a :write_schema option. This will
inform the RDF translator to write out the BEL RDFS.

The RDFS_SCHEMA was set as frozen Array constant.

refs #95