parse all BEL relationships in BEL Script
With this change the BEL Script parser will recognize all BEL
relationships. The semantic validity of using these BEL relationship
is a different matter.

Take the BEL Statement "A actsIn activity(A)". This relationship follows
as a derivation of "activity(A)", given BEL derivation rules,  though if
captured as a BEL Nanopub one could provide citation, context, and
further metadata.