added streaming combiner option; refactored names
The previous StreamingEvidenceCombiner combined based on apriori
reference maps for annotations and namespaces. This has been renamed to
MapReferencesCombiner.

A streaming combiner has been added that takes the annotation and
namespace references from the first evidence in the stream. This should
only be used when the references are consistent amongst the evidence
collection being translated.